perf: implement route-based code splitting and vendor chunking (closes #4)
The single 767 KB JS bundle (232 KB gzip) was causing Vite's chunk-size
warning and slowing initial page load, especially on mobile.

- Convert all five page routes to TanStack Router's .lazy.tsx pattern so
  each page (dashboard, categories, transactions, login, register) is a
  separate dynamically-imported chunk
- Add build.rolldownOptions.output.manualChunks in vite.config.ts to
  split node_modules into stable, cacheable vendor chunks: vendor-react
  (React + ReactDOM), vendor-query (TanStack Query), vendor-recharts
  (Recharts + D3 deps), and a catch-all vendor chunk
- No single chunk now exceeds 500 KB; initial load drops to ~141 KB gzip
  and recharts (74 KB gzip) only loads when the dashboard is visited

Closes #4
