+++ src/routes/_auth/login.tsx
@@ -1,90 +1,3 @@
-import { useMutation } from '@tanstack/react-query'
-import { Link, createFileRoute, useNavigate } from '@tanstack/react-router'
-import { useState } from 'react'
-import { Button } from '@/components/ui/button'
-import { Input } from '@/components/ui/input'
-import { loginUser } from '@budget-buddy-org/budget-buddy-contracts'
-import { useAuthStore } from '@/stores/auth.store'
-import type { AuthToken, LoginRequest } from '@budget-buddy-org/budget-buddy-contracts'
+import { createFileRoute } from '@tanstack/react-router'
 
-export const Route = createFileRoute('/_auth/login')({
-  component: LoginPage,
-})
-
-function LoginPage() {
-  const navigate = useNavigate()
-  const setAuth = useAuthStore((s) => s.setAuth)
-  const [username, setUsername] = useState('')
-  const [password, setPassword] = useState('')
-
-  const login = useMutation({
-    mutationFn: async (body: LoginRequest) => {
-      const { data } = await loginUser({ body })
-      return data as AuthToken
-    },
-    onSuccess: (data) => {
-      setAuth(data.access_token, data.refresh_token)
-      navigate({ to: '/' })
-    },
-  })
-
-  return (
-    <div className="space-y-4">
-      <div>
-        <h2 className="text-lg font-medium">Sign in</h2>
-        <p className="text-sm text-muted-foreground">Enter your credentials to continue</p>
-      </div>
-
-      <form
-        onSubmit={(e) => {
-          e.preventDefault()
-          login.mutate({ username, password })
-        }}
-        className="space-y-3"
-      >
-        <div className="space-y-1">
-          <label className="text-sm font-medium" htmlFor="username">
-            Username
-          </label>
-          <Input
-            id="username"
-            type="text"
-            value={username}
-            onChange={(e) => setUsername(e.target.value)}
-            autoComplete="username"
-            required
-          />
-        </div>
-
-        <div className="space-y-1">
-          <label className="text-sm font-medium" htmlFor="password">
-            Password
-          </label>
-          <Input
-            id="password"
-            type="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            autoComplete="current-password"
-            required
-          />
-        </div>
-
-        {login.isError && (
-          <p className="text-sm text-destructive">Invalid credentials. Please try again.</p>
-        )}
-
-        <Button type="submit" className="w-full" disabled={login.isPending}>
-          {login.isPending ? 'Signing in…' : 'Sign in'}
-        </Button>
-      </form>
-
-      <p className="text-center text-sm text-muted-foreground">
-        Don&apos;t have an account?{' '}
-        <Link to="/register" className="font-medium text-foreground underline-offset-4 hover:underline">
-          Sign up
-        </Link>
-      </p>
-    </div>
-  )
-}
+export const Route = createFileRoute('/_auth/login')({})
